requirements.txt
# Core dependencies
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
pydantic>=2.10.0,<3.0
sqlalchemy==2.0.36
alembic==1.14.0

# Database - Modern PostgreSQL driver with async support
psycopg[binary]>=3.1,<4.0
prometheus-fastapi-instrumentator>=7.0,<8.0
chromadb==0.5.23

# Document processing
PyPDF2==3.0.1
pdfplumber==0.11.4
camelot-py==0.11.0
tabula-py==2.10.0
pytesseract==0.3.13
pillow==11.0.0
python-magic==0.4.27
pandas==2.2.3
openpyxl==3.1.5
lxml==5.3.0
ofxtools==0.9.5
python-dateutil==2.9.0.post0
pytz==2024.2
python-docx==1.1.2

# AI/ML
openai==1.56.2
tiktoken==0.8.0
sentence-transformers==3.3.1

# File monitoring
watchdog==6.0.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.0
openai-agents==0.2.11
tenacity==9.0.0
structlog==24.4.0
rich==13.9.4
click==8.1.7
pyyaml==6.0.2

# Web interface
streamlit==1.41.0
plotly==5.24.1
streamlit-tree-select>=0.1.0
streamlit-aggrid>=1.0.0
streamlit-elements>=0.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0